% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_cis.R
\name{ci_change}
\alias{ci_change}
\title{Make a y axis for an epicurve, prepackaged with defaults}
\usage{
ci_change(
  new_val,
  old_val,
  dist = c("poisson", "negbin"),
  output = c("pct_change", "ratio"),
  conf = 0.95,
  theta
)
}
\arguments{
\item{new_val}{The integer value more recently occurring in time}

\item{old_val}{The integer value less recently occurring in time}

\item{dist}{the assumed distribution the two values are taken from. choose from "poisson" and "negbin"}

\item{output}{should the output be via the confidence interval of
the ratio between the two values or the percentage change. choose from
"pct_change" or "ratio"}

\item{conf}{confidence level, defaults to 0.95}

\item{theta}{theta parameter, for negative binomial distribution only}

\item{expand}{the expand argument to \code{scale_y_continuous()}. Defaults to \code{expansion(mult = c(0, NA))}}
}
\description{
Make a y axis for an epicurve, prepackaged with defaults
}
\author{
Henry LS, Finlay Campbell
}
