% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_country_names.R
\name{convert_country_names}
\alias{convert_country_names}
\title{function to convert country names via a synonym table, using a standard to/from structure}
\usage{
convert_country_names(
  x,
  country_col = "country",
  show_corrections = TRUE,
  ref = country_synonyms,
  ref_from_col = "from",
  ref_to_col = "to"
)
}
\arguments{
\item{x}{\itemize{
\item the dataframe where conversion takes place
}}

\item{country_col}{\itemize{
\item a character object with the name of the column in your dataset corresponding to country.
Defaults to "country"
}}

\item{show_corrections}{\itemize{
\item a logical. Should a logical column (named is_corrected) be additionally outputted, showing
where corrections have taken place?
}}

\item{ref}{a data frame. A reference synonym dataframe, usually with a "to" and "from" column. Defaults to the prepackaged country_synonyms}

\item{ref_from_col}{\itemize{
\item a character object with the name of the column with non-official synonyms in the reference dataset
}}

\item{from_to_col}{\itemize{
\item a character object with the name of the column with official country names in the reference dataset
}}
}
\description{
function to convert country names via a synonym table, using a standard to/from structure
}
\author{
Henry LS
}
