% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving_average.R
\name{moving_average}
\alias{moving_average}
\title{function to apply a moving average to a numeric column. Use this function within \code{mutate()}
Note that function requires data to be sorted by date. By default calculates a 7 day moving average.}
\usage{
moving_average(val, date_val, before = 3, after = 3)
}
\arguments{
\item{val}{\itemize{
\item which \code{numeric} class variable to be used
}}

\item{date_val}{\itemize{
\item which \code{date} class index to be used
}}

\item{before}{\itemize{
\item the number of days to include before index. Defaults to 3.
}}

\item{after}{\itemize{
\item the number of days to include after index. Defaults to 3.
}}
}
\description{
function to apply a moving average to a numeric column. Use this function within \code{mutate()}
Note that function requires data to be sorted by date. By default calculates a 7 day moving average.
}
