% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{official_pal}
\alias{official_pal}
\title{official_pal}
\format{
A tibble with 89 rows and 5 variables:
\describe{
\item{\code{type}}{type of pallette}
\item{\code{subtype}}{subtype of pallette}
\item{\code{tint}}{the degree to which the original colour has been tinted (made lighter)}
\item{\code{name}}{the name of the assoicated colour}
\item{\code{hex}}{the hex colour of the associated colour}
}
}
\usage{
official_pal
}
\description{
table of colours defined according to different schemas in the WHO style guide
}
\keyword{datasets}
