% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perc_change.R
\name{perc_change}
\alias{perc_change}
\title{This function calculates percentage change between two values (i.e. for week
on week change)}
\usage{
perc_change(new_val, old_val)
}
\arguments{
\item{new_val}{the value more recent in time}

\item{old_val}{the value further back in time}
}
\description{
This function calculates percentage change between two values (i.e. for week
on week change)
}
\author{
HLS
}
