% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_country}
\alias{ref_country}
\title{ref_country}
\format{
A tibble with 238 rows and 4 variables:
\describe{
\item{\code{country}}{character the UN name of the country}
\item{\code{who_region}}{character the WHO region of the country}
\item{\code{iso2}}{character The ISO2 code for the country}
\item{\code{iso3}}{character The ISO3 code for the country}
}
}
\usage{
ref_country
}
\description{
a reference table of all UN countries, alongside WHO region, and iso2 and 3 codes.
}
\keyword{datasets}
