% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel_region.R
\name{relabel_region}
\alias{relabel_region}
\title{Change region labels from long to short form and vice versa.}
\usage{
relabel_region(region)
}
\arguments{
\item{region}{A character vector of regions to relabel.}
}
\description{
Change region labels from long to short form and vice versa.
}
\author{
Finlay Campbell, Henry LS, Steve Kerr
}
