% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel_who.R
\name{relabel_who}
\alias{relabel_who}
\title{Change WHO region labels from long to short form and from region to office.}
\usage{
relabel_who(
  label,
  output = c("keep", "region", "office"),
  length = c("switch", "short", "long")
)
}
\arguments{
\item{label}{A character vector of regions in long or short form, as office
or region.}

\item{output}{Should the output be provided as a region ("region"), office
("office") or kept as initially specified ("keep").}

\item{length}{Should the output be provided in abbreviated form ("short"), long
form ("long") or switched ("switch"). Defaults to "switch".}
}
\description{
Change WHO region labels from long to short form and from region to office.
}
\examples{
## return the associated region as long text
relabel_region(
  c("EMR", "AFRO", "Regional Office for the Americas"),
  output = "region",
  length = "long"
)

## convert long format to short format and vice versa
relabel_region(c("EMR", "AFRO", "Regional Office for the Americas"))

}
\author{
Finlay Campbell, Henry LS, Steve Kerr
}
