% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_y_epicurve.R
\name{scale_y_epicurve}
\alias{scale_y_epicurve}
\title{Make a y axis for an epicurve, prepackaged with defaults}
\usage{
scale_y_epicurve(
  mode = c("cases", "incidence"),
  limits = c(0, NA),
  lower_maximum = 5,
  expand = expansion(mult = c(0, NA)),
  ...
)
}
\arguments{
\item{mode}{one of "cases" or "incidence" - if in case mode, only integer values will be shown in labels,
otherwise, there is potential to show decimal points}

\item{limits}{The limits argument to \code{scale_y_contiunous()}.
Y-axis limits for the epicurve - by default this is \code{c(0, NA)} - i.e. default for the upper limit,
and 0 as the lower limit}

\item{lower_maximum}{The lowest possible upper limit, as a numeric value. This will mean the upper y axis limit, for all facets will by this
number at minimum. Useful when dealing with very low numbers, defaults to 5.}

\item{expand}{the expand argument to \code{scale_y_continuous()}. Defaults to \code{expansion(mult = c(0, NA))}}
}
\description{
Make a y axis for an epicurve, prepackaged with defaults
}
\author{
Henry LS
#'
}
