% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_style.R
\name{text_style}
\alias{text_style}
\title{This function creates HTML text styling via inputted css,, for use in rmarkdown documents}
\usage{
text_style(
  text,
  size = "1em",
  col = "black",
  weight = "normal",
  style = "normal",
  ...
)
}
\arguments{
\item{text}{the text within an rmarkdown document to be stylised}

\item{size}{the size of the text using css conventions.
Use a number followed by 'px' (pixels) or 'em' (relative measure to current font size)}

\item{col}{the colour of the font using css conventions.
Use a standard HTML colour name ("red", "blue", "grey"), rgb colour ("rgb(255, 99, 71)"),
or a hex colour ("#e41a1c")}

\item{weight}{the weight of the font using css conventions.
Use an number value (normal is 100), or a css naming convention (e.g. "light", "normal", "bold")}

\item{style}{the style of the font using css conventions.
Use a css naming convention (e.g. "normal", "italic")}

\item{...}{any additional css arguments to apply to the text, provided in css style
e.g. ('font-family: "Times New Roman"'). These can be separated by ';', or provided as a vector.}
}
\description{
This function creates HTML text styling via inputted css,, for use in rmarkdown documents
}
\author{
HLS
}
