% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_who.R
\name{theme_who}
\alias{theme_who}
\title{a general ggplot2 theme for use across products}
\usage{
theme_who(base_size = 11, ...)
}
\arguments{
\item{base_size}{the base size of text in the plot}

\item{...}{additional theme arguments}
}
\description{
a general ggplot2 theme for use across products
}
\author{
HLS
}
