% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_who_gt.R
\name{theme_who_gt}
\alias{theme_who_gt}
\title{A general table theme for the \code{{gt}} packag}
\usage{
theme_who_gt(x, mode = c("light", "dark"))
}
\arguments{
\item{x}{a gt object}
}
\description{
A general table theme for the \code{{gt}} packag
}
\author{
HLS
}
