% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_agg_epicurve.R
\name{vis_agg_epicurve}
\alias{vis_agg_epicurve}
\title{This function visualizes epidemic curves based on aggregated case or death counts.}
\usage{
vis_agg_epicurve(
  dat,
  start_date = NULL,
  end_date = Sys.Date(),
  var = new_cases,
  date_var = date_reported,
  by_week = FALSE,
  plot_title = NULL,
  fill_by = "who_region_long",
  facet_by = NULL,
  free_y = TRUE,
  col_pal = outbreak_pal("qual1"),
  facet_cols = 2,
  return_data = FALSE
)
}
\arguments{
\item{dat}{A data frame containing number of cases/deaths per day to be visualized.}

\item{start_date}{The start date of the visualization. Defaults to NULL, which means that the earliest date in the data will be used.}

\item{end_date}{The end date of the visualization. Defaults to \code{Sys.Date()}, which means that the current date will be used.}

\item{var}{A variable to be used for the y-axis. Defaults to \code{new_cases}.}

\item{date_var}{A variable to be used for the x-axis. Defaults to \code{date_reported}.}

\item{by_week}{A logical value indicating whether to aggregate the data by week. Defaults to FALSE.}

\item{plot_title}{The plot title. Defaults to NULL, which means that no title will be used.}

\item{fill_by}{The variable to be used for the fill aesthetic. Defaults to "who_region_long".}

\item{facet_by}{The variable to be used for faceting. Defaults to NULL, which means that no faceting will occur.}

\item{free_y}{A logical value indicating whether the y-axis should be scaled independently for each facet. Defaults to TRUE.}

\item{col_pal}{The color palette to be used. Defaults to \code{outbreak_pal("qual1")}.}

\item{facet_cols}{The number of columns in the facet grid. Defaults to 2.}

\item{return_data}{A logical value indicating whether to return the data once processed. Defaults to FALSE.}
}
\value{
If return_data is TRUE, returns a data frame. Otherwise, it returns a ggplot object.
}
\description{
This function visualizes epidemic curves based on aggregated case or death counts.
}
\author{
Henry Laurenson-Schafer, Steve Kerr
}
