% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who_palette.R
\name{who_palette}
\alias{who_palette}
\title{This function exports a range of predefined palettes}
\usage{
who_palette(
  type = c("qual1", "qual2", "qual3", "qual4", "binary1", "binary2", "who", "region",
    "change1", "change2", "official"),
  labels = NULL,
  n = Inf
)
}
\arguments{
\item{type}{a \code{character} name of a palette to be exported. Choose from "qual1", "qual2", "qual3", "qual4", "binary1", "binary2",
"who", "region", "change1", or "change2"}

\item{labels}{optional - a \code{character} vector of labels to apply to each colour.}

\item{n}{optional - the number of colours to output. By default, the maximum in the palette is used.
Most useful for change palettes, where the difference between colour grades will increase as n decreases.}
}
\description{
This function exports a range of predefined palettes
}
\author{
HLS
}
