% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/official_pal.R
\name{who_palette_official}
\alias{who_palette_official}
\title{This function exports a range of predefined palettes}
\usage{
who_palette_official(
  type = c("primary", "secondary", "ramp", "grey", "diverging", "qualitative"),
  subtype = NULL,
  tint = NULL,
  n = NULL,
  use_names = FALSE
)
}
\arguments{
\item{type}{a \code{character} name of a palette to be exported. Choose from "primary", "secondary", "ramp", "grey", "diverging","qualitative"}

\item{subtype}{optional - a character defining a subtype of palette, for use in secondary, grey, and ramp palettes - this confines the palette
to a specific hue. Where no subtype has been defined, ramp and secondary palettes, will not be confined to a single hue.}

\item{tint}{optional - an integer value between 10 and 100 defining the degree of tint applied to ramp or secondary palettes. Use this to obtain darker or lighter shades
of colours. Integer value snaps to nearest closest available tint. Tints greater than 100 are possible for ramp palettes.}

\item{n}{optional - the number of colours to output. Here, the colours with the most spacing possible are chosen, while accounting for
colourblindness.}

\item{use_names}{should the output vector have the colour names added? Defaults to FALSE.}
}
\description{
This function exports a range of predefined palettes
}
\author{
HLS
}
