% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whotools-package.R
\docType{package}
\name{whotools-package}
\alias{whotools}
\alias{whotools-package}
\title{whotools: R Tools for WHO Analytics}
\description{
This package contains low-level tools for conducting WHO analytics in R.
}
\author{
\strong{Maintainer}: Finlay Campbell \email{campellf@who.int}

Authors:
\itemize{
  \item Henry Laurenson-Schafer \email{laurensonschaferh@who.int}
  \item Steve Kerr \email{kerrs@who.int}
}

}
\keyword{internal}
