% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_synonyms}
\alias{country_synonyms}
\title{country_synonyms}
\format{
A tibble with 39 rows and 2 variables:
\describe{
\item{\code{from}}{character the non-official synonym}
\item{\code{to_name}}{character the official country name}
\item{\code{to_iso3}}{character the iso3 code of the country}
}
}
\usage{
country_synonyms
}
\description{
synonym table of common country synoyms that are \emph{NOT} WHO official
}
\keyword{datasets}
