% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_who_gt.R
\name{theme_who_gt}
\alias{theme_who_gt}
\title{a general table theme for the \code{{gt}} packag}
\usage{
theme_who_gt(x)
}
\arguments{
\item{x}{a gt object}
}
\description{
a general table theme for the \code{{gt}} packag
}
\author{
HLS
}
